requirements.txt
gradio>=4.0.0
google-generativeai>=0.3.0
google-genai>=0.3.0
numpy>=1.24.0
scikit-learn>=1.3.0